Add regression tests
